% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomac-transform.R
\name{derive_year_sheet_ind}
\alias{derive_year_sheet_ind}
\title{Determine Sheet Prefix for Firepred Data Dictionary}
\usage{
derive_year_sheet_ind(ind_perims, year, pre_pst_split_ind)
}
\arguments{
\item{ind_perims}{(logical): Boolean indicator. If TRUE, the file is a perimeter file,
otherwise it is an incident report.}

\item{pre_pst_split_ind}{(logical): Boolean indicator. If TRUE, then variable names for perimeters
change before and after 2015 (as before 11/7/2019 changes),
so that the `*_pre15` sheets are used in the Firepred
data dictionary for before 2015 and the `*_pst15' for the
years afterwards. If FALSE (default since 11/7/2019 changes),
the `*_pst15' sheets are used across the years for perimeters.}

\item{yrs}{(numeric): Vector with years which data need to be transformed}
}
\value{
A vector of `pre15` or `pst15` according to which Data Dictionary sheet should be considered.
}
\description{
Before 11/7/2019, both fire perimeters and incident reports displayed
two sets of variable, depending on whether the data were recorded before
or after 2015.
As of 11/7/2019, fire perimeters have been updated so that the variable
names are consistent across years now. However, incident reports still
presents two different variable names depending on whether the year
was before or after 2015.
This function looks at which sets of variable names to look at based on
whether it's a perimeter or incident report file, which year the file is
and whether we are dealing with data before 11//7/2019 changes.
}
