% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtbs-transform.R
\name{format_transform_mtbs}
\alias{format_transform_mtbs}
\title{Format Specific MTBS Shapefile}
\usage{
format_transform_mtbs(dl_date, mtbs_type, orig_colnames, tform_colnames,
  ds_source = "mtbs")
}
\arguments{
\item{dl_date}{(date) : Date in which the file were downloaded. This is
going to look for a folder in the `data/{ds_source}/` named
as this date}

\item{mtbs_type}{(character) : This specifies the type of MTBS data we want
to import and takes 2 values \code{mtbs_perimeter_data, mtbs_fod_pts_data}}

\item{orig_colnames}{(array) : Name of the columns in the original shapefile}

\item{tform_colnames}{(array) : Name of the columns in the transformed shapefile}

\item{ds_source}{(character) : data names. Default to "mtbs"}
}
\value{
A sf data frame containing the shapefile contents
}
\description{
Format Specific MTBS Shapefile
}
\examples{
\dontrun{
list_colnames <- mtbs_transform_col_names(mtbs_type = "mtbs_perimeter_data")
format_transform_mtbs(ds_source = "mtbs",
                    dl_date = base::as.Date("2019-07-06"),
                    mtbs_type = "mtbs_perimeter_data",
                    orig_colnames = list_colnames$ORIG_COLNAMES,
                    tform_colnames = list_colnames$TFORM_COLNAMES)
}
}
