% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpa-fod-extract.R
\name{fpa_data_download}
\alias{fpa_data_download}
\title{Download data for FPA data}
\usage{
fpa_data_download(fpa_fod_all, data_folder = "data", remove = FALSE)
}
\arguments{
\item{data_folder}{(character) : "data" (default) data folder name}

\item{remove}{(logical) : FALSE (default) remove .zip file after extraction}

\item{noaa_all}{(tibble) : Data and Metadata info}
}
\description{
Download data for FPA data
}
