% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpa-fod-extract.R
\name{fpa_metadata}
\alias{fpa_metadata}
\title{Download full data and metadata information for FPA data}
\usage{
fpa_metadata(fpa_fod_urls)
}
\arguments{
\item{fpa_fod_urls}{(tibble) : FPA URLs tibble}
}
\description{
Download full data and metadata information for FPA data
}
