% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpa-fod-extract.R
\name{fpa_read_data}
\alias{fpa_read_data}
\title{Extracting information on the download URLs for FPA data}
\usage{
fpa_read_data()
}
\description{
Extracting information on the download URLs for FPA data
}
