% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpa-fod-extract.R
\name{fpa_setup_variables}
\alias{fpa_setup_variables}
\title{Setup the global variables for FPA data}
\usage{
fpa_setup_variables(dat_flag = "fpa_fod", data_dl_method = "libcurl",
  metadata_dirname = "metadata", url = NULL)
}
\arguments{
\item{dat_flag}{(character) : "fpa_fod" (default) name of the data - will
appear as data folder name when data are saved}

\item{data_dl_method}{(character) : "libcurl" (default) data download method}

\item{metadata_dirname}{(character) : "metadata" (default) metadata
folder name}

\item{url}{(character) : NULL (default) data url - the default one is encoded
in the function body}
}
\description{
Setup the global variables for FPA data
}
