% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpa-fod-transform.R
\name{fpa_transform}
\alias{fpa_transform}
\title{Wrapper for FPA FOD transformation pipeline.}
\usage{
fpa_transform(dl_date, ds_source = "fpa_fod", fpa_fod_type = "GPKG")
}
\arguments{
\item{dl_date}{(date) : Date in which the file were downloaded. This is
going to look for a folder in the `data/{ds_source}/` named
as this date}

\item{ds_source}{(character) : data names. Default to "fpa"}

\item{fpa_fod_type}{(character) : The specifics FPA FOD data format.
Currently only supported option is "GPKG", which imports the geopackage
data format}
}
\value{
(list) : A list of transformed shapefiles, in which the names is the year
}
\description{
Wrapper for FPA FOD transformation pipeline.
}
