% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomac-extract.R
\name{geomac_data_download}
\alias{geomac_data_download}
\title{Download data for geomac data}
\usage{
geomac_data_download(geomac_all, year_periods = 2016:2018,
  data_folder = "data", remove = FALSE)
}
\arguments{
\item{year_periods}{(numeric) : 1800:1800 (default) array of years to be
downloaded}

\item{data_folder}{(character) : "data" (default) data folder name}

\item{remove}{(logical) : FALSE (default) remove .zip file after extraction}

\item{noaa_all}{(tibble) : Data and Metadata info}
}
\description{
Download data for geomac data
}
