% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomac-extract.R
\name{geomac_extract}
\alias{geomac_extract}
\title{Wrapper for GEOMAC data extraction pipeline.
This function will setup global variables, gather information and urls on
data and metadata and download the data.}
\usage{
geomac_extract(year_periods = 2016:2018, data_folder = "data",
  dat_flag = "geomac", data_dl_method = "libcurl", url = NULL,
  remove = FALSE, metadata_dirname = "metadata")
}
\arguments{
\item{year_periods}{(numeric) : 1800:1800 (default) array of years to be
downloaded}

\item{data_folder}{(character) : "data" (default) data folder name}

\item{dat_flag}{(character) : "geomac" (default) name of the data -
will appear as data folder name when data are
saved}

\item{data_dl_method}{(character) : "libcurl" (default) data download method}

\item{url}{(character) : NULL (default) data url - the default one is encoded
in the function body}

\item{remove}{(logical) : FALSE (default) remove .zip file after extraction}

\item{metadata_dirname}{(character) : "metadata" (default) metadata
folder name}
}
\description{
Wrapper for GEOMAC data extraction pipeline.
This function will setup global variables, gather information and urls on
data and metadata and download the data.
}
