% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomac-transform.R
\name{geomac_format_shapefiles}
\alias{geomac_format_shapefiles}
\title{Format GEOMAC shapefiles given path}
\usage{
geomac_format_shapefiles(shapefiles_path, geomac_type)
}
\arguments{
\item{shapefiles_path}{(character) : Path of the shapefile to be transformed}

\item{geomac_type}{: Can be either `geomac_fire_perimeters` or
`geomac_situation_report`. Different types filter
on different columns for the creation of the shapefile}
}
\value{
A shapefile object
}
\description{
Format GEOMAC shapefiles given path
}
