% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomac-extract.R
\name{geomac_metadata}
\alias{geomac_metadata}
\title{Download full data and metadata information for GEOMAC data}
\usage{
geomac_metadata(geomac_urls)
}
\arguments{
\item{geomac_urls}{(tibble) : geomac data information}
}
\description{
Download full data and metadata information for GEOMAC data
}
