% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomac-extract.R
\name{geomac_read_data}
\alias{geomac_read_data}
\title{Extracting information on the metadata for GEOMAC data}
\usage{
geomac_read_data()
}
\description{
Extracting information on the metadata for GEOMAC data
}
