% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl-utils.R
\name{get_common_cols}
\alias{get_common_cols}
\title{Gets the intersection or union of columns in a list of dataframes
This is used in the \code{geomac} and \code{mtbs} load code}
\usage{
get_common_cols(inp_df_list, set_group_fn)
}
\arguments{
\item{inp_df_list}{(list) : A list of data frames which may have some columns
in common}

\item{set_group_fn}{(function) : Typically \code{intersect} or \code{union}.
Will determine if we get the common intersection or union of all column
names across all the data frames in the list}
}
\description{
Gets the intersection or union of columns in a list of dataframes
This is used in the \code{geomac} and \code{mtbs} load code
}
