% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl-utils.R
\name{get_ddict}
\alias{get_ddict}
\title{Get data dictionary names from single Excel sheet}
\usage{
get_ddict(ddict_xlsx_path, sheet_nm, cols_rng)
}
\arguments{
\item{ddict_xlsx_path}{: .xlsx file path}

\item{sheet_nm}{: sheet number}

\item{cols_rng}{: column range in the sheet to be selected}
}
\value{
tibble with data dictionary names
}
\description{
Get data dictionary names from single Excel sheet
}
