% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomac-transform.R
\name{get_ddict_geomac}
\alias{get_ddict_geomac}
\title{Get data dictionaries for column names given .xlxs file}
\usage{
get_ddict_geomac(ddict_xlsx_path, cols_rng)
}
\arguments{
\item{ddict_xlsx_path}{: .xlxs file path}

\item{cols_rng}{: Column range in each of the excel sheets name}
}
\value{
List with column names in the different .xlxs pages
}
\description{
Get data dictionaries for column names given .xlxs file
}
