% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomac-transform.R
\name{get_geomac_mtda_paths}
\alias{get_geomac_mtda_paths}
\title{Provides a helper dataframe to list all of the files extracted for geomac
for a specified year of download}
\usage{
get_geomac_mtda_paths(ds_source, dl_date, yr)
}
\arguments{
\item{ds_source}{(character) :}

\item{dl_date}{(date) :}

\item{yr}{(integer) : The specifics geomac year shapefile we want to
transform}
}
\value{
A tibble of metadata paths
}
\description{
Provides a helper dataframe to list all of the files extracted for geomac
for a specified year of download
}
