% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomac-transform.R
\name{get_geomac_new_cnames}
\alias{get_geomac_new_cnames}
\title{Set Shapefile column names given dictionary names fetched from .xlsx
data dictionary file.}
\usage{
get_geomac_new_cnames(
  fpath,
  year,
  ddict_type,
  ddict_xlsx_path,
  cols_rng,
  target_srid = 4326
)
}
\arguments{
\item{fpath}{: shapefile path}

\item{year}{: year the shapefile is referred to}

\item{ddict_type}{: dictionary type - this needs to be corresponding to one
of the data dictionaries retrieved in the .xlsx}

\item{ddict_xlsx_path}{: path of the data dictionary .xlsx}

\item{cols_rng}{: column ranges to extract from the .xlsx}

\item{target_srid}{(numeric): Transform the geometries into this coordinate
system, specified by SRID (default: 4326, GPS latitude/longitude)}
}
\value{
Shapefile dataframe
}
\description{
Set Shapefile column names given dictionary names fetched from .xlsx
data dictionary file.
}
