% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn-load-utils.R
\name{get_ghcn_stations_us_unq}
\alias{get_ghcn_stations_us_unq}
\title{Get the unique United States station from our \code{ghcn stations} table}
\usage{
get_ghcn_stations_us_unq(conn)
}
\arguments{
\item{conn}{(connection) : A connection to the PostGIS database containing
the \code{ghcn_stations} table}
}
\value{
(tibble): A tibble of unique United States GHCN stations with
the associated station id, elevation, location fields
}
\description{
Get the unique United States station from our \code{ghcn stations} table
}
