% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtbs-transform.R
\name{get_mtbs_mtda_paths}
\alias{get_mtbs_mtda_paths}
\title{Provides a helper dataframe to list all of the files extracted for MTBS
for a specified year of download}
\usage{
get_mtbs_mtda_paths(dl_date, mtbs_type, ds_source = "mtbs")
}
\arguments{
\item{dl_date}{(date) : Date in which the file were downloaded. This is
going to look for a folder in the `data/{ds_source}/` named
as this date}

\item{mtbs_type}{(character) : This specifies the type of MTBS data we want
to import and takes 2 values \code{mtbs_perimeter_data, mtbs_fod_pts_data}}

\item{ds_source}{(character) : data names. Default to "mtbs"}
}
\value{
A tibble of filenames
}
\description{
Provides a helper dataframe to list all of the files extracted for MTBS
for a specified year of download
}
\examples{
\dontrun{
get_mtbs_mtda_paths(ds_source = "mtbs",
                    dl_date = base::as.Date("2019-07-06"),
                    mtbs_type = "mtbs_perimeter_data")
}
}
