% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms-noaa-transform.R
\name{get_noaa_stormevents_mtda_paths}
\alias{get_noaa_stormevents_mtda_paths}
\title{Provides a helper dataframe to list all of the files extracted for NOAA STORMEVENTS
for a specified year of download}
\usage{
get_noaa_stormevents_mtda_paths(dl_date, yr,
  ds_source = "noaa_stormevents")
}
\arguments{
\item{dl_date}{(date) : Date in which the file were downloaded. This is
going to look for a folder in the `data/{ds_source}/` named
as this date}

\item{yr}{(integer) : The specifics NOAA STORMEVENTS year shapefile we want to
transform}

\item{ds_source}{(character) : name of the data source, default to "noaa_stormevents"}
}
\value{
A tibble with filenames
}
\description{
Provides a helper dataframe to list all of the files extracted for NOAA STORMEVENTS
for a specified year of download
}
