% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms-noaa-transform.R
\name{get_noaa_stormevents_new_cnames}
\alias{get_noaa_stormevents_new_cnames}
\title{Set Shapefile column names given dictionary names fetched from .xlsx
data dictionary file.}
\usage{
get_noaa_stormevents_new_cnames(fpath, year, noaa_type, ddict_xlsx_path,
  cols_rng)
}
\arguments{
\item{fpath}{: shapefile path}

\item{year}{: year the shapefile is referred to}

\item{ddict_xlsx_path}{: path of the data dictionary .xlsx}

\item{cols_rng}{: column ranges to extract from the .xlsx}

\item{ddict_type}{: dictionary type - this needs to be corresponding to one
of the data dictionaries retrieved in the .xlsx}
}
\value{
Shapefile dataframe
}
\description{
Set Shapefile column names given dictionary names fetched from .xlsx
data dictionary file.
}
