% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdi-transform.R
\name{get_noaa_swdi_mtda_paths}
\alias{get_noaa_swdi_mtda_paths}
\title{Provides a helper dataframe to list all of the files extracted for NOAA SWDI
for a specified year of download}
\usage{
get_noaa_swdi_mtda_paths(
  ds_source = "noaa_swdi",
  dl_date,
  yr,
  noaa_swdi_ind_tiles,
  noaa_swdi_type
)
}
\arguments{
\item{ds_source}{(character) : name of the data source, default to "noaa_swdi"}

\item{dl_date}{(date) : Date in which the file were downloaded. This is
going to look for a folder in the `data/{ds_source}/` named
as this date}

\item{yr}{(integer) : The specifics NOAA SWDI year CSV we want to
transform}

\item{noaa_swdi_ind_tiles}{(integer) : A value in \code{{0, 1}}, with 1
indicating that tiles data is to be downloaded and extracted,
and 0 indicating non-tiles data is to be downloaded and extracted. Currently
only a value of 1 i.e. tiles is supported.}

\item{noaa_swdi_type}{(character) : Specifying the type of NOAA-SWDI data to
download and extract. Can be one of the following values:
\code{hail, mda, meso, nldn, plsr, structure, tvs, warn}}
}
\value{
A tibble with filenames
}
\description{
Provides a helper dataframe to list all of the files extracted for NOAA SWDI
for a specified year of download
}
