% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl-utils.R
\name{get_sel_cols}
\alias{get_sel_cols}
\title{Selects the specified columns from the input data frame
This is used in the \code{geomac} and \code{mtbs} load code}
\usage{
get_sel_cols(inp_df, sel_cols)
}
\arguments{
\item{inp_df}{(data.frame) : The input data frame}

\item{sel_cols}{(character) : A character of column names we want to
\code{select} from the data frame}
}
\description{
Selects the specified columns from the input data frame
This is used in the \code{geomac} and \code{mtbs} load code
}
