% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl-utils.R
\name{get_subset_cols}
\alias{get_subset_cols}
\title{Selects the specified columns from a list of input data frames.
This is used in the \code{geomac} and \code{mtbs} load code}
\usage{
get_subset_cols(inp_df_list, sel_cols)
}
\arguments{
\item{inp_df_list}{(list) : A list of data frames which may have some columns
in common}

\item{sel_cols}{(character) : A character of column names we want to
\code{select} from the data frame}
}
\description{
Selects the specified columns from a list of input data frames.
This is used in the \code{geomac} and \code{mtbs} load code
}
