% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpa-fod-transform.R
\name{get_transform_fpa}
\alias{get_transform_fpa}
\title{Transform the FPA FOD fires file}
\usage{
get_transform_fpa(fpath)
}
\arguments{
\item{fpath}{(character) : The path to the FPA_FOD GPKG file we want
to transform. The column names will get converted to lower case and spaces
replaced by underscores via the \code{janitor} package}
}
\value{
(sf object) : transformed shapefile
}
\description{
Transform the FPA FOD fires file
}
