% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn-transform.R
\name{get_transform_ghcn}
\alias{get_transform_ghcn}
\title{Transform the GHCN file}
\usage{
get_transform_ghcn(fpath, new_colnames)
}
\arguments{
\item{fpath}{(character) : The path to the shapefile we want to transform}

\item{new_colnames}{(character) : New list of colnames we want to set for
our transformed output shapefile. If \code{NULL} the column names will get
converted to lower case and spaces replaced by underscores via the
\code{janitor} package}
}
\value{
(sf object) : transformed shapefile
}
\description{
TODO: Need to add assertions to check the dimension of the column names and
      also the values. Somehow need to parse in the metadata from the web
      and get the field values in an automated manner.
}
