% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn-stations-transform.R
\name{get_transform_ghcnd_stations}
\alias{get_transform_ghcnd_stations}
\title{Transformed GHCN-Daily Stations file. This is metadata for the GHCN-daily
file. It contains additional info on US/Canada states and country
information for each station}
\usage{
get_transform_ghcnd_stations()
}
\value{
(tibble) : Transformed tibble of GHCN stations metadata
}
\description{
Transformed GHCN-Daily Stations file. This is metadata for the GHCN-daily
file. It contains additional info on US/Canada states and country
information for each station
}
\examples{
\dontrun{
ghcnd_stations_tform <- backburner::get_transform_ghcnd_stations()
}

}
