% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtbs-transform.R
\name{get_transform_mtbs}
\alias{get_transform_mtbs}
\title{Transform the MTBS file}
\usage{
get_transform_mtbs(fpath, exp_orig_colnames, new_colnames)
}
\arguments{
\item{fpath}{(character) : The path to the shapefile we want to transform}

\item{new_colnames}{(character) : New list of colnames we want to set for
our transformed output shapefile. If \code{NULL} the column names will get
converted to lower case and spaces replaced by underscores via the
\code{janitor} package}
}
\value{
(sf object) : transformed shapefile
}
\description{
Transform the MTBS file
}
