% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn-transform.R
\name{ghcn_transform}
\alias{ghcn_transform}
\title{Wrapper for GHCN transformation pipeline.}
\usage{
ghcn_transform(dl_date, yrs, ds_source = "noaa_ghcn_daily")
}
\arguments{
\item{dl_date}{(date) : Date in which the file were downloaded. This is
going to look for a folder in the `data/{ds_source}/` named
as this date}

\item{yrs}{(array): numerical array with the years which needs to be transformed}

\item{ds_source}{(character) : data names. Default to "noaa_ghcn_daily"}
}
\value{
(list) : A list of transformed shapefiles, in which the names is the year
}
\description{
Wrapper for GHCN transformation pipeline.
}
