% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn-transform.R
\name{ghcn_transform_col_names}
\alias{ghcn_transform_col_names}
\title{Setup the column names for GHCN.}
\usage{
ghcn_transform_col_names()
}
\value{
(array): character array with the variable names
}
\description{
Setup the column names for GHCN.
}
