% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl-utils.R
\name{ind_re_match}
\alias{ind_re_match}
\title{Indicator function to return a \code{{0, 1}} integer value if there is a
valid regex match for the given input string}
\usage{
ind_re_match(string, pattern)
}
\arguments{
\item{string}{(character) : Input string to match regex against
(must be a single element vector)}

\item{pattern}{(character) : regex string pattern}
}
\description{
Indicator function to return a \code{{0, 1}} integer value if there is a
valid regex match for the given input string
}
