% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtbs-extract.R
\name{mtbs_data_download}
\alias{mtbs_data_download}
\title{Download data for MTBS data}
\usage{
mtbs_data_download(mtbs_all, data_folder = "data", remove = FALSE)
}
\arguments{
\item{data_folder}{(character) : "data" (default) data folder name}

\item{remove}{(logical) : FALSE (default) remove .zip file after extraction}

\item{noaa_all}{(tibble) : Data and Metadata info}
}
\description{
Download data for MTBS data
}
