% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtbs-extract.R
\name{mtbs_extract}
\alias{mtbs_extract}
\title{Wrapper for MTBS data extraction pipeline.
This function will setup global variables, gather information and urls on
data and metadata and download the data.}
\usage{
mtbs_extract(dat_flag = "mtbs", data_dl_method = "libcurl",
  url = NULL, remove = FALSE, data_folder = "data",
  metadata_dirname = "metadata")
}
\arguments{
\item{dat_flag}{(character) : "mtbs" (default) name of the data -
will appear as data folder name when data are
saved}

\item{data_dl_method}{(character) : "libcurl" (default) data download method}

\item{url}{(character) : NULL (default) data url - the default one is encoded
in the function body}

\item{remove}{(logical) : FALSE (default) remove .zip file after extraction}

\item{data_folder}{(character) : "data" (default) data folder name}

\item{metadata_dirname}{(character) : "metadata" (default) metadata
folder name}
}
\description{
Wrapper for MTBS data extraction pipeline.
This function will setup global variables, gather information and urls on
data and metadata and download the data.
}
