% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtbs-extract.R
\name{mtbs_metadata}
\alias{mtbs_metadata}
\title{Download full data and metadata information for MTBS data}
\usage{
mtbs_metadata(mtbs_mtda)
}
\arguments{
\item{mtbs_mtda}{(tibble) : MTBS url and data information tibble}
}
\description{
Download full data and metadata information for MTBS data
}
