% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtbs-extract.R
\name{mtbs_read_data}
\alias{mtbs_read_data}
\title{Extracting information on the download URLs for MTBS functions}
\usage{
mtbs_read_data()
}
\description{
Extracting information on the download URLs for MTBS functions
}
