% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtbs-transform.R
\name{mtbs_transform}
\alias{mtbs_transform}
\title{Wrapper for MTBS transformation pipeline.
This function will transform both mtbs files, i.e.
\code{mtbs_perimeter_data, mtbs_fod_pts_data}.}
\usage{
mtbs_transform(dl_date, ds_source = "mtbs")
}
\arguments{
\item{dl_date}{(date) : Date in which the file were downloaded. This is
going to look for a folder in the `data/{ds_source}/` named
as this date}

\item{ds_source}{(character) : data names. Default to "mtbs"}
}
\value{
A list with two shapefiles for loading, in which the key is the
        mtbs_type
}
\description{
Wrapper for MTBS transformation pipeline.
This function will transform both mtbs files, i.e.
\code{mtbs_perimeter_data, mtbs_fod_pts_data}.
}
