% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn-extract.R
\name{noaa_ghcn_data_download}
\alias{noaa_ghcn_data_download}
\title{Download data for NOOA GHCN data}
\usage{
noaa_ghcn_data_download(noaa_all, year_periods = 1800:1800,
  data_folder = "data", remove = FALSE)
}
\arguments{
\item{noaa_all}{(tibble) : Data and Metadata info}

\item{year_periods}{(numeric) : 1800:1800 (default) array of years to be
downloaded}

\item{data_folder}{(character) : "data" (default) data folder name}

\item{remove}{(logical) : FALSE (default) remove .zip file after extraction}
}
\description{
Download data for NOOA GHCN data
}
