% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms-noaa-extract.R
\name{noaa_stormevents_extract}
\alias{noaa_stormevents_extract}
\title{Wrapper for NOOA STORMEVENTS CSV data extraction pipeline.
This function will setup global variables, gather information and urls on
data and metadata and download the data.}
\usage{
noaa_stormevents_extract(
  year_periods = 1950:1950,
  data_folder = "data",
  dat_flag = "noaa_stormevents",
  remove = FALSE,
  data_dl_method = "libcurl",
  url = NULL,
  metadata_dirname = "metadata"
)
}
\arguments{
\item{year_periods}{(numeric) : 1950:1950 (default) array of years to be
downloaded}

\item{data_folder}{(character) : "data" (default) data folder name}

\item{dat_flag}{(character) : "noaa_stormevents" (default) name of the data -
will appear as data folder name when data are
saved}

\item{remove}{(logical) : FALSE (default) remove .zip file after extraction}

\item{data_dl_method}{(character) : "libcurl" (default) data download method}

\item{url}{(character) : NULL (default) data url - the default one is encoded
in the function body}

\item{metadata_dirname}{(character) : "metadata" (default) metadata
folder name}
}
\description{
Wrapper for NOOA STORMEVENTS CSV data extraction pipeline.
This function will setup global variables, gather information and urls on
data and metadata and download the data.
}
\examples{
\dontrun{
noaa_stormevents_extract(year_periods = 1800:1801)
}
}
