% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms-noaa-transform.R
\name{noaa_stormevents_identify}
\alias{noaa_stormevents_identify}
\title{Identify NOOA type from file name}
\usage{
noaa_stormevents_identify(path, types = c("details", "locations",
  "fatalities"))
}
\description{
Identify NOOA type from file name
}
