% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms-noaa-extract.R
\name{noaa_stormevents_metadata}
\alias{noaa_stormevents_metadata}
\title{Download full data and metadata information for NOOA STORMEVENTS CSV data}
\usage{
noaa_stormevents_metadata(noaa_urls)
}
\arguments{
\item{noaa_urls}{(tibble) : NOOA STORMEVENTS CSV URLs tibble}
}
\description{
Download full data and metadata information for NOOA STORMEVENTS CSV data
}
