% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms-noaa-transform.R
\name{noaa_stormevents_parse_datetime}
\alias{noaa_stormevents_parse_datetime}
\title{Convert NOAA timestamps to R datetimes.}
\usage{
noaa_stormevents_parse_datetime(datetimes, tzs, format)
}
\description{
NOAA inconveniently uses a weird timestamp format in different time zones,
and R's strptime can't accept vectorized input with one time zone per element
(instead of one fixed time zone). So we have to manually loop.
}
