% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms-noaa-extract.R
\name{noaa_stormevents_read_data}
\alias{noaa_stormevents_read_data}
\title{Extracting information on the download URLs for NOOA STORMEVENTS CSV data}
\usage{
noaa_stormevents_read_data()
}
\description{
Extracting information on the download URLs for NOOA STORMEVENTS CSV data
}
