% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storms-noaa-transform.R
\name{noaa_stormevents_transform}
\alias{noaa_stormevents_transform}
\title{Wrapper for NOAA STORMEVENTS transformation pipeline.}
\usage{
noaa_stormevents_transform(dl_date, yrs, ds_source = "noaa_stormevents")
}
\arguments{
\item{dl_date}{(date) : Date in which the file were downloaded. This is
going to look for a folder in the `data/{ds_source}/` named
as this date}

\item{yrs}{(array): numerical array with the years which needs to be transformed}

\item{ds_source}{(character) : data names. Default to "noaa_stormevents"}
}
\value{
(list) : A list of transformed shapefiles, in which the names is the year
}
\description{
Wrapper for NOAA STORMEVENTS transformation pipeline.
}
