% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdi-extract.R
\name{noaa_swdi_extract}
\alias{noaa_swdi_extract}
\title{Wrapper for NOAA GCHN data extraction pipeline.
This function will setup global variables, gather information and urls on
data and metadata and download the data.}
\usage{
noaa_swdi_extract(
  year_periods = 1995:1995,
  noaa_swdi_ind_tiles,
  noaa_swdi_type,
  data_folder = "data",
  dat_flag = "noaa_swdi",
  remove = FALSE,
  data_dl_method = "libcurl",
  url = NULL,
  metadata_dirname = "metadata"
)
}
\arguments{
\item{year_periods}{(numeric) : 1800:1800 (default) array of years to be
downloaded}

\item{noaa_swdi_ind_tiles}{(integer) : A value in \code{{0, 1}}, with 1
indicating that tiles data is to be downloaded and extracted,
and 0 indicating non-tiles data is to be downloaded and extracted}

\item{noaa_swdi_type}{(character) : Specifying the type of NOAA-SWDI data to
download and extract. Can be one of the following values:
\code{hail, mda, meso, nldn, plsr, structure, tvs, warn}}

\item{data_folder}{(character) : "data" (default) data folder name}

\item{dat_flag}{(character) : "noaa_swdi_daily" (default) name of the data -
will appear as data folder name when data are
saved}

\item{remove}{(logical) : FALSE (default) remove .zip file after extraction}

\item{data_dl_method}{(character) : "libcurl" (default) data download method}

\item{url}{(character) : NULL (default) data url - the default one is encoded
in the function body}

\item{metadata_dirname}{(character) : "metadata" (default) metadata
folder name}
}
\description{
Wrapper for NOAA GCHN data extraction pipeline.
This function will setup global variables, gather information and urls on
data and metadata and download the data.
}
\examples{
\dontrun{
noaa_swdi_extract(year_periods = 1995:1996,
                  noaa_swdi_ind_tiles = 1,
                  noaa_swdi_type = "hail")
}
}
