% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdi-extract.R
\name{noaa_swdi_metadata}
\alias{noaa_swdi_metadata}
\title{Download full data and metadata information for NOAA SWDI data}
\usage{
noaa_swdi_metadata(noaa_urls, noaa_swdi_ind_tiles)
}
\arguments{
\item{noaa_urls}{(tibble) : NOAA SWDI URLs tibble}

\item{noaa_swdi_ind_tiles}{(integer) : A value in \code{{0, 1}}, with 1
indicating that tiles data is to be downloaded and extracted,
and 0 indicating non-tiles data is to be downloaded and extracted}
}
\description{
Download full data and metadata information for NOAA SWDI data
}
