% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdi-extract.R
\name{noaa_swdi_read_data}
\alias{noaa_swdi_read_data}
\title{Extracting information on the download URLs for NOAA SWDI data}
\usage{
noaa_swdi_read_data()
}
\description{
Extracting information on the download URLs for NOAA SWDI data
}
