% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn-extract.R
\name{nooa_ghcn_metadata}
\alias{nooa_ghcn_metadata}
\title{Download full data and metadata information for NOOA GHCN data}
\usage{
nooa_ghcn_metadata(noaa_urls)
}
\arguments{
\item{noaa_urls}{(tibble) : NOOA GHCN URLs tibble}
}
\description{
Download full data and metadata information for NOOA GHCN data
}
