% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn-extract.R
\name{nooa_ghcn_read_data}
\alias{nooa_ghcn_read_data}
\title{Extracting information on the download URLs forNOOA GHCN data}
\usage{
nooa_ghcn_read_data()
}
\description{
Extracting information on the download URLs forNOOA GHCN data
}
