% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn-extract.R
\name{nooa_ghcn_setup_variables}
\alias{nooa_ghcn_setup_variables}
\title{Setup the global variables for NOOA GHCN data}
\usage{
nooa_ghcn_setup_variables(
  dat_flag = "noaa_ghcn_daily",
  data_dl_method = "libcurl",
  metadata_dirname = "metadata",
  url = NULL
)
}
\arguments{
\item{dat_flag}{(character) : "noaa_ghcn_daily" (default) name of the data -
will appear as data folder name when data are
saved}

\item{data_dl_method}{(character) : "libcurl" (default) data download method}

\item{metadata_dirname}{(character) : "metadata" (default) metadata
folder name}

\item{url}{(character) : NULL (default) data url - the default one is encoded
in the function body}
}
\description{
Setup the global variables for NOOA GHCN data
}
