% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdi-extract.R
\name{nooa_swdi_metadata}
\alias{nooa_swdi_metadata}
\title{Download full data and metadata information for NOOA GHCN data}
\usage{
nooa_swdi_metadata(noaa_urls)
}
\arguments{
\item{noaa_urls}{(tibble) : NOOA GHCN URLs tibble}
}
\description{
Download full data and metadata information for NOOA GHCN data
}
