% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdi-transform.R
\name{nooa_swdi_transform_col_names}
\alias{nooa_swdi_transform_col_names}
\title{Setup the column names for NOAA SWDI}
\usage{
nooa_swdi_transform_col_names()
}
\value{
(array): character array with the variable names
}
\description{
Setup the column names for NOAA SWDI
}
