% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl-utils.R
\name{sp_transform_sf}
\alias{sp_transform_sf}
\title{Transform \code{sp} shapefile to an \code{sf} shapefile object with
a given \code{crs} coordinate reference system. The column names are
converted to lowercase and non-trailing spaces replaced with underscores
via the \code{janitor::clean_names} function}
\usage{
sp_transform_sf(sp, crs = 4326)
}
\arguments{
\item{sp}{(sp) : The spatial polygon \code{sp} object which you want to
transform into an \code{sf} object}

\item{crs}{(sf) : The Coordinate Reference System (CRS) system with
default value of 4326}
}
\value{
(sf) : The \code{sf} converted object
}
\description{
Transform \code{sp} shapefile to an \code{sf} shapefile object with
a given \code{crs} coordinate reference system. The column names are
converted to lowercase and non-trailing spaces replaced with underscores
via the \code{janitor::clean_names} function
}
