% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris-transform.R
\name{tigris_calc_areas}
\alias{tigris_calc_areas}
\title{Calculates areas of tigris state, county, and military geometry (polygons)
in units of square meters, square kilometers, square miles. It appends
these columns onto the input tigris tibble
(of either state, county, or military).}
\usage{
tigris_calc_areas(tigris_tbl)
}
\arguments{
\item{tigris_tbl}{(sf) : Processed tigris \code{sf} \code{tibble} already
projected to CRS 4326}
}
\value{
(tibble): The input tigris \code{sf} \code{tibble} with
3 additional columns area calculations appended to it.
}
\description{
Calculates areas of tigris state, county, and military geometry (polygons)
in units of square meters, square kilometers, square miles. It appends
these columns onto the input tigris tibble
(of either state, county, or military).
}
