% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris-extract.R
\name{tigris_extract}
\alias{tigris_extract}
\title{Extracts the TIGER/Line Files and Shapefiles for US States, US Counties,
and US military lines for a given year, and resolution, in preparation
for transform purposes}
\usage{
tigris_extract(tig_res = "5m", tig_year = 2019, tig_cb = FALSE, tig_crs = 4326)
}
\arguments{
\item{tig_res}{(character) : The resolution of the cartographic boundary
file (if \code{cb == TRUE}). Defaults to '500k'; options include
'5m' (1:5 million) and '20m' (1:20 million). Default is '5m'}

\item{tig_year}{(integer) : The year for which you want to download the}

\item{tig_cb}{(logical) : If cb is set to \code{TRUE}, download a
generalized (1:500k) states file. Defaults to \code{FALSE} (the most
detailed TIGER/Line file)}

\item{tig_crs}{(integer) : The Coordinate Reference System (CRS) system with
default value of 4326}
}
\value{
(list) : A list of shapefiles with cleaned up column names for
US States, US Counties, and US military lines for a given year, and
resolution, with cleaned column names, ready for modeling transformations.
}
\description{
Extracts the TIGER/Line Files and Shapefiles for US States, US Counties,
and US military lines for a given year, and resolution, in preparation
for transform purposes
}
\examples{
\dontrun{
tigris_out <- tigris_extract(tig_res = '5m', tig_year = 2019,
                            tig_cb = FALSE, tig_crs = 4326)
}
}
