% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire-base-180d-sql-qry-utils.R
\name{wrap_gen_query_fire_base_full_feat_180d}
\alias{wrap_gen_query_fire_base_full_feat_180d}
\title{Wrapper function to generate a single combined or separate \code{SQL} files
for all fire features joined in the mtbs table for modeling the ground
process and fire size. This joins all of the lagged aggregated time summaries
mapped to individual MTBS fires by specified spatial degrees over a 180 day
lag period for each fire}
\usage{
wrap_gen_query_fire_base_full_feat_180d(
  fire_base_start_date = base::as.Date("2000-01-01"),
  max_lag_fire_base_days = 180,
  spat_deg_rad = 0.5,
  summary_type = "AVG",
  out_qry_dir,
  ind_comb_qry = TRUE
)
}
\arguments{
\item{fire_base_start_date}{(date) : The minimum date to filter MTBS fire data, we
will i.e. dataset will only contain MTBS fires after this date}

\item{max_lag_fire_base_days}{(integer) : The number of days for NOAA SWDI weather
data to lag behind the \code{fire_base_start_date}}

\item{spat_deg_rad}{(double) : The spatial degree radius to consider around
each fire when joining weather data (default value is 0.5)}

\item{summary_type}{(character) : The \code{SQL} summary function to apply
to the GHCN-D weather data e.g. "AVG", "MEDIAN" (default value is "AVG")}

\item{out_qry_dir}{(character) : The directory to save the query. A full
file path will be generated as part of the output}

\item{ind_comb_qry}{(logical) : If \code{TRUE} then produces a single
combined \code{sql} query for all features combined. If \code{FALSE}
then produces a separate \code{sql} queries for individual GHCN-D feature}
}
\value{
(tibble) : A single or separate \code{SQL} files for single
GHCN features joined to MTBS fire_ids by specified spatial and temporal
parameters
}
\description{
Wrapper function to generate a single combined or separate \code{SQL} files
for all fire features joined in the mtbs table for modeling the ground
process and fire size. This joins all of the lagged aggregated time summaries
mapped to individual MTBS fires by specified spatial degrees over a 180 day
lag period for each fire
}
\examples{
\dontrun{
library("tidyverse")
wrap_gen_query_fire_base_full_feat_180d(fire_base_start_date =
                                      base::as.Date("2000-01-01"),
                                    max_lag_fire_base_days = 180,
                                    spat_deg_rad = 0.5,
                                    summary_type = "AVG",
                                    out_qry_dir = ".",
                                    ind_comb_qry = TRUE)
}
}
