% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcn-base-sql-qry-utils.R
\name{wrap_gen_query_ghcn_base_sfeat}
\alias{wrap_gen_query_ghcn_base_sfeat}
\title{Wrapper function to generate a single combined or separate \code{SQL} files
for individual feature NOAA SWDI lagged aggregated time summaries mapped to
individual MTBS fires by specified spatial degrees over a 180 day lag period
for each fire}
\usage{
wrap_gen_query_ghcn_base_sfeat(
  sfeat_types,
  out_qry_dir,
  out_tbl_name_pfx,
  fire_base_tbl_name,
  ind_comb_qry = TRUE
)
}
\arguments{
\item{out_qry_dir}{(character) : The directory to save the query. A full
file path will be generated as part of the output}

\item{out_tbl_name_pfx}{(character) : The default table name prefix to take.
This will define prefix the name of the output table and also the output
query file name. In this case a suitable default value is
"ghcnd_observations"}

\item{fire_base_tbl_name}{(character) : The table name for the GHCN-D
observations table. In this case a suitable default value is
"ghcnd_observations"}

\item{ind_comb_qry}{(logical) : If \code{TRUE} then produces a single
combined \code{sql} query for all features combined. If \code{FALSE}
then produces a separate \code{sql} queries for individual GHCN-D feature}

\item{sfeat}{(character) : The NOAA GHCN features to create a separate table
for e.g \code{c("PRCP", "SNOW", "SNWD", "TMAX", "TMIN", "TOBS", "TAVG")}.
In this wrapper the user should pass in a vector of all the features that
are contained in the GHCN-D table. This will generate combined or separate
queries to generate an individual table for each such GHCN-D feature}
}
\value{
(tibble) : A single or separate \code{SQL} files for single
GHCN features joined to MTBS fire_ids by specified spatial and temporal
parameters
}
\description{
Wrapper function to generate a single combined or separate \code{SQL} files
for individual feature NOAA SWDI lagged aggregated time summaries mapped to
individual MTBS fires by specified spatial degrees over a 180 day lag period
for each fire
}
\examples{
\dontrun{
library("tidyverse")
wrap_gen_query_ghcn_base_sfeat(sfeat = c("PRCP", "SNOW", "SNWD", "TMAX", "TMIN", "TOBS", "TAVG"),
                               out_qry_dir = ".",
                               out_tbl_name_pfx = "ghcnd_observations",
                               fire_base_tbl_name = "ghcnd_observations",
                               ind_comb_qry = TRUE)
}
}
