% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomac-transform.R
\name{geomac_transform}
\alias{geomac_transform}
\title{Full Pipeline for Transforming GEOMAC Shapefiles}
\usage{
geomac_transform(dl_date, yrs, ds_source = "geomac")
}
\arguments{
\item{dl_date}{(date) : Date in which the file were downloaded. This is
going to look for a folder in the `data/{ds_source}/` named
as this date}

\item{yrs}{(numeric): Vector with years which data need to be transformed}

\item{ds_source}{(character) : data names. Default to "geomac"}
}
\value{
A list with two shapefiles for loading, in which the key is the
        mtbs_type
}
\description{
This function will transform both geomac files, i.e.
\code{geomac_fire_perimeters, geomac_situation_report}.
}
